requirements.txt
djangorestframework
Django==1.11.1
olefile==0.44
Pillow==4.2.0
pytz==2017.2
